+++ src/main/java/com/facturation/model/Produit.java
@@ -25,7 +25,7 @@ public class Produit implements Serializable {
 
     private String code;
 
-    private BigDecimal prix;
+    private Double prix;
 
     private int stock;
